analysis new intervention - multiple selects to step 1
