+++ client/src/components/badge/component.tsx
@@ -3,7 +3,14 @@ import { useCallback } from 'react';
 
 import type { BadgeProps } from './types';
 
-const Badge: React.FC<BadgeProps> = ({ data, children, className, removable = false, onClick }) => {
+const Badge: React.FC<BadgeProps> = ({
+  data,
+  children,
+  className,
+  removable = false,
+  onClick,
+  background = true,
+}) => {
   const handleClick = useCallback(
     (e) => {
       e.preventDefault();
@@ -16,8 +23,9 @@ const Badge: React.FC<BadgeProps> = ({ data, children, className, removable = fa
   return (
     <span
       className={classNames(
-        'inline-flex items-center py-0.5 rounded-full font-medium bg-green-100 text-green-700',
+        'inline-flex items-center py-0.5 font-medium  text-green-700',
         removable ? 'pl-2 pr-0.5' : 'px-2',
+        !!background && 'bg-green-100 rounded-full',
         className,
       )}
     >
