+++ client/src/components/select/component.tsx
@@ -20,6 +20,7 @@ const ScenariosComparison: React.FC<SelectProps> = (props: SelectProps) => {
     showSearch = false,
     disabled = false,
     label,
+    id,
     options = [],
     current = options[0],
     loading = false,
