+++ client/src/components/tree-select/primary-layout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
