+++ client/src/components/tree-select/secondary-layout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
