+++ client/src/components/tree-select/secondary-layout/types.d.ts
@@ -9,6 +9,7 @@ export type TreeSelectOptions = TreeSelectOption[];
 export type TreeSelectProps = {
   maxBadges?: number;
   placeholder?: string;
+  currentOptions: TreeSelectOption[];
   searchPlaceholder?: string;
   showSearch?: boolean;
   multiple?: boolean;
@@ -17,5 +18,8 @@ export type TreeSelectProps = {
   current: TreeSelectOptions;
   onChange?: (selected: SelectOption) => unknown;
   onSearch?: (query: string) => unknown;
+  isOpen: boolean;
   theme?: 'primary' | 'secondary';
+  handleOpen: Dispatch<SetStateAction<boolean>>;
+  handleRemoveBadget: (option: any) => void;
 };
