+++ client/src/store/features/analysis/index.ts
@@ -26,7 +26,10 @@ export type AnalysisState = {
     materials: Option[];
     origins: Option[];
     suppliers: Option[];
-    interventionType: 'Source from a new supplier or location' | 'Change production efficiency' | 'Switch to a new material';
+    interventionType:
+      | 'Source from a new supplier or location'
+      | 'Change production efficiency'
+      | 'Switch to a new material';
   };
 };
